Move plugins to osb package
