+++ api/osb/check_instance_owner_plugin.go
@@ -1,4 +1,4 @@
-package plugins
+package osb
 
 import (
 	"fmt"
