+++ api/osb/check_platform_id_plugin.go
@@ -1,4 +1,4 @@
-package plugins
+package osb
 
 import (
 	"fmt"
